+++ CHANGES.md
@@ -1,3 +1,9 @@
+# Version 10.1.0
+
+- Add an option to allow switching the death saves from blind to private
+- Hide the status of the death saves on the character sheet (supports the vanilla and tidy5e sheets)
+- Add an explanation via chat message about the blind death save for the rolling player
+
 # Version 10.0.0
 
 - Initial release for Foundry VTT v10